rename load/store instruction lookups
