+++ jolt-core/src/host/mod.rs
@@ -24,12 +24,10 @@ use crate::{
     field::JoltField,
     jolt::{
         instruction::{
-            div::DIVInstruction, divu::DIVUInstruction, mulh::MULHInstruction,
+            div::DIVInstruction, divu::DIVUInstruction, lb::LBInstruction, lbu::LBUInstruction,
+            lh::LHInstruction, lhu::LHUInstruction, mulh::MULHInstruction,
             mulhsu::MULHSUInstruction, rem::REMInstruction, remu::REMUInstruction,
-            virtual_lb::VirtualLBInstruction, virtual_lbu::VirtualLBUInstruction,
-            virtual_lh::VirtualLHInstruction, virtual_lhu::VirtualLHUInstruction,
-            virtual_sb::VirtualSBInstruction, virtual_sh::VirtualSHInstruction,
-            VirtualInstructionSequence,
+            sb::SBInstruction, sh::SHInstruction, VirtualInstructionSequence,
         },
         vm::{bytecode::BytecodeRow, rv32i_vm::RV32I, JoltTraceStep},
     },
@@ -199,12 +197,12 @@ impl Program {
                 tracer::RV32IM::DIVU => DIVUInstruction::<32>::virtual_trace(row),
                 tracer::RV32IM::REM => REMInstruction::<32>::virtual_trace(row),
                 tracer::RV32IM::REMU => REMUInstruction::<32>::virtual_trace(row),
-                tracer::RV32IM::SH => VirtualSHInstruction::<32>::virtual_trace(row),
-                tracer::RV32IM::SB => VirtualSBInstruction::<32>::virtual_trace(row),
-                tracer::RV32IM::LBU => VirtualLBUInstruction::<32>::virtual_trace(row),
-                tracer::RV32IM::LHU => VirtualLHUInstruction::<32>::virtual_trace(row),
-                tracer::RV32IM::LB => VirtualLBInstruction::<32>::virtual_trace(row),
-                tracer::RV32IM::LH => VirtualLHInstruction::<32>::virtual_trace(row),
+                tracer::RV32IM::SH => SHInstruction::<32>::virtual_trace(row),
+                tracer::RV32IM::SB => SBInstruction::<32>::virtual_trace(row),
+                tracer::RV32IM::LBU => LBUInstruction::<32>::virtual_trace(row),
+                tracer::RV32IM::LHU => LHUInstruction::<32>::virtual_trace(row),
+                tracer::RV32IM::LB => LBInstruction::<32>::virtual_trace(row),
+                tracer::RV32IM::LH => LHInstruction::<32>::virtual_trace(row),
                 _ => vec![row],
             })
             .map(|row| {
