+++ jolt-core/src/jolt/instruction/mod.rs
@@ -155,6 +155,10 @@ pub mod bgeu;
 pub mod bne;
 pub mod div;
 pub mod divu;
+pub mod lb;
+pub mod lbu;
+pub mod lh;
+pub mod lhu;
 pub mod mul;
 pub mod mulh;
 pub mod mulhsu;
@@ -163,6 +167,8 @@ pub mod mulu;
 pub mod or;
 pub mod rem;
 pub mod remu;
+pub mod sb;
+pub mod sh;
 pub mod sll;
 pub mod slt;
 pub mod sltu;
@@ -175,14 +181,8 @@ pub mod virtual_assert_lte;
 pub mod virtual_assert_valid_div0;
 pub mod virtual_assert_valid_signed_remainder;
 pub mod virtual_assert_valid_unsigned_remainder;
-pub mod virtual_lb;
-pub mod virtual_lbu;
-pub mod virtual_lh;
-pub mod virtual_lhu;
 pub mod virtual_move;
 pub mod virtual_movsign;
-pub mod virtual_sb;
-pub mod virtual_sh;
 pub mod xor;
 
 #[cfg(test)]
